+++ config.js
@@ -5,6 +5,7 @@ var config = {};
 config.twilioAccountSid = process.env.TWILIO_ACCOUNT_SID;
 config.twilioAuthToken  = process.env.TWILIO_AUTH_TOKEN;
 config.twilioPhoneNumber= process.env.TWILIO_PHONE_NUMBER;
+
 console.log(config);
 
 var requiredConfig = [config.twilioAccountSid,
