+++ constants.js
@@ -1,0 +1,3 @@
+var constants = {};
+
+constants.uploadDir = 'uploads';
