+++ public/index.html
@@ -36,13 +36,6 @@
           <li ng-repeat="visit in visits"> {{visit}} </li>
         </ul>
       </div>
-      <div class="text-visits">
-        <form novalidate class="message-form">
-            Message (English): <input type="text" ng-model="msg.messageEng" /><br/>
-            Message (Spanish): <input type="text" ng-model="msg.messageSpan" /><br/>
-            <input type="submit" ng-click="update(msg)" value="Send Messages" />
-        </form>
-      </div>
     </div>
 
     <script src="//ajax.googleapis.com/ajax/libs/angularjs/1.5.8/angular.min.js" type="text/javascript"></script>
