+++ server.js
@@ -10,19 +10,14 @@ var twilio = require('twilio');
 //Not vendor - but I am my own product
 var config = require('./config');
 var port = process.env.PORT || 3000;
-var upload = require('./routes/uploads')
+var upload = require('./routes/upload')
+var uploads = require('./routes/uploads')
 var app = express();
 
 // Node config
 app.use(express.static('public'));
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({extended : true}));
-
-
-// App config
-const uploadDir = 'uploads';
-const twilioClient = new twilio.RestClient(config.twilioAccountSid,
-  config.twilioAuthToken);
 
 
 //Routes
