--- test/1.csv
@@ -1,1 +1,0 @@
-Latoya, Williams, 3365098391, Date_Placeholder
